requirements.txt
openpyxl>=3.0.0
tkinterdnd2>=0.3.0
Pillow>=9.0.0
